+++ tests/test_dataset.py
@@ -333,6 +333,92 @@ class TestDataSet(unittest.TestCase):
         for record in self.dataset:
             self.assertIsInstance(record['tags'], list)
 
+    def test_create_index(self):
+        self.dataset.create_index(name='address_state_index', keys=['address.state'])
+
+        self.assertEqual(self.dataset.indexes['address_state_index']['keys'], ['address.state'])
+        self.assertEqual(list(self.dataset.indexes['address_state_index']['values'].keys()), ['CA', 'TX', 'NY'])
+
+        for index_definition, records in self.dataset.indexes['address_state_index']['values'].items():
+            self.assertEqual(len(records), 1)
+            self.assertEqual(index_definition, records[0].walk('address.state'))
+
+    def test_drop_index(self):
+        self.dataset.create_index(name='address_state_index', keys=['address.state'])
+        self.dataset.drop_index('address_state_index')
+
+        self.assertNotIn('address_state_index', self.dataset.indexes)
+
+    def test_find_index(self):
+        self.dataset.create_index(name='address_state_index', keys=['address.state'])
+        index = self.dataset.find_index(keys=['address.state'])
+
+        self.assertEqual(index, 'address_state_index')
+
+    def test_refresh_index(self):
+        self.dataset.create_index(name='address_state_index', keys=['address.state'])
+
+        self.dataset[0].assign('address.state', 'TX')
+        self.dataset.refresh_index('address_state_index')
+
+        self.assertEqual(len(self.dataset.indexes['address_state_index']['values']), 2)
+        self.assertEqual(list(self.dataset.indexes['address_state_index']['values'].keys()), ['TX', 'NY'])
+
+    def test_inner_join(self):
+        righthand_dataset = DataSet([
+            {
+                'name': 'John Doe',
+                'dob': '1990-01-01',
+                'medications': ['med-1, med-2']
+            },
+            {
+                'name': 'Jane Smith',
+                'dob': '1985-05-15',
+                'medications': ['med-3']
+            },
+            # intentionally not included for inner join test
+            # {
+            #     'name': 'Jane Smith',
+            #     'dob': '1992-07-20',
+            # }
+
+        ])
+
+        self.dataset.join(righthand_dataset, left_keys=['name', 'dob'], right_keys=['name', 'dob'], inner=True)
+
+        self.assertEqual(len(self.dataset), 2)
+        for record in self.dataset:
+            self.assertIn('medications', record)
+            self.assertIsInstance(record['medications'], list)
+            self.assertEqual(record['medications'], ['med-1, med-2'] if record['name'] == 'John Doe' else ['med-3'])
+
+    def test_join(self):
+        righthand_dataset = DataSet([
+            {
+                'name': 'John Doe',
+                'dob': '1990-01-01',
+                'medications': ['med-1, med-2']
+            },
+            {
+                'name': 'Jane Smith',
+                'dob': '1985-05-15',
+                'medications': ['med-3']
+            },
+            # intentionally not included for inner join test
+            # {
+            #     'name': 'Jane Smith',
+            #     'dob': '1992-07-20',
+            # }
+
+        ])
+
+        self.dataset.join(righthand_dataset, left_keys=['name', 'dob'], right_keys=['name', 'dob'], inner=False)
+
+        self.assertEqual(len(self.dataset), 3)
+        self.assertEqual(self.dataset[0]['medications'], ['med-1, med-2'])
+        self.assertEqual(self.dataset[1]['medications'], ['med-3'])
+        self.assertNotIn('medications', self.dataset)
+
 class TestWalkableDict(unittest.TestCase):
     def setUp(self):